Move flaky stuff to own module
